refactor, dedup
